can: isotp: use unique skbcnt for CF echo verification
The cfecho tag is used to doublecheck echo'ed consecutive frames (CF).
Instead of grabbing four bytes from the CAN frame data better make use of
the unique skbcnt value to reduce complexity for the upcoming CAN XL
ISO 15765-2 support.

can_set_skb_uid(skb) now creates a unique skbcnt value in the receive path
in can_receive() and in isotp.c in the tx path for FC frames.
If the skbcnt is already set in isotp.c it will not be changed at rx time.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>